==> notebooks/03_dashboard_prep.ipynb <==
# app/db.py
from sqlalchemy import create_engine
import pandas as pd

engine = create_engine(
    "postgresql+psycopg2://postgres:110868@localhost:5432/geolife"
)

pd.read_sql("SELECT 1;", engine)


query = """
SELECT table_schema, table_name
FROM information_schema.tables
WHERE table_schema = 'mart';
"""

pd.read_sql(query, engine)


user_df = pd.read_sql(
    "SELECT * FROM mart.user_mobility_summary;",
    engine
)

user_df.head()


user_df[["total_trips", "active_days", "avg_idle_time_ratio"]].head()


user_df.sort_values(
    "total_trips",
    ascending=False
).head(5)[
    ["user_id", "total_trips", "active_days", "avg_idle_time_ratio"]
]


trip_df = pd.read_sql(
    "SELECT * FROM mart.trip_features;",
    engine
)

trip_df.head()


trip_df[[
    "total_distance_km",
    "trip_duration_min",
    "avg_speed_kmph",
    "idle_time_ratio"
]].describe()


sample_user = user_df.iloc[0]["user_id"]

sample_user


trip_df[trip_df["user_id"] == sample_user].shape


user_df[user_df["user_id"] == sample_user][
    ["total_trips", "active_days", "avg_trip_distance_km"]
]


trip_df[trip_df["user_id"] == sample_user][
    ["total_distance_km", "trip_duration_min"]
].sum()


anomaly_df = pd.read_sql(
    "SELECT * FROM mart.trip_anomalies;",
    engine
)

anomaly_df.head()


anomaly_df["anomaly_score"].value_counts().sort_index()


(
    (anomaly_df["anomaly_score"] > 0).mean() * 100
)


anomaly_df[
    anomaly_df["anomaly_score"] >= 2
][[
    "user_id",
    "trip_id",
    "total_distance_km",
    "trip_duration_min",
    "avg_speed_kmph",
    "idle_time_ratio",
    "night_trip_flag",
    "anomaly_score"
]].head(10)


anomaly_df.groupby("user_id")["anomaly_score"].mean().sort_values(ascending=False).head()


user_dash_df = pd.read_sql(
    "SELECT * FROM mart.user_mobility_summary;",
    engine
)

anomaly_dash_df = pd.read_sql(
    "SELECT * FROM mart.trip_anomalies;",
    engine
)


user_dash_df = user_dash_df[
    [
        "user_id",
        "total_trips",
        "active_days",
        "avg_trip_distance_km",
        "avg_speed_kmph",
        "avg_idle_time_ratio",
        "weekday_trip_ratio",
        "night_trip_ratio"
    ]
]


anomaly_dash_df = anomaly_dash_df[
    [
        "user_id",
        "trip_id",
        "total_distance_km",
        "trip_duration_min",
        "avg_speed_kmph",
        "idle_time_ratio",
        "night_trip_flag",
        "anomaly_score"
    ]
]


user_dash_df.head()


anomaly_dash_df.head()
